refactor(cli): improve testability by moving logic to consoleOutput module
